+++ reclass/adapters/salt.py
@@ -3,15 +3,17 @@
 #
 # This file is part of reclass (http://github.com/madduck/reclass)
 #
-# Copyright © 2007–13 martin f. krafft <madduck@madduck.net>
+# Copyright © 2007–14 martin f. krafft <madduck@madduck.net>
 # Released under the terms of the Artistic Licence 2.0
 #
 
 import os, sys, posix
 
-from reclass import get_nodeinfo, get_inventory, output
+from reclass import get_storage, output
+from reclass.core import Core
 from reclass.errors import ReclassException
-from reclass.config import find_and_read_configfile, get_options
+from reclass.config import find_and_read_configfile, get_options, \
+        path_mangler
 from reclass.constants import MODE_NODEINFO
 from reclass.defaults import *
 from reclass.version import *
@@ -21,14 +23,25 @@ def ext_pillar(minion_id, pillar,
                inventory_base_uri=OPT_INVENTORY_BASE_URI,
                nodes_uri=OPT_NODES_URI,
                classes_uri=OPT_CLASSES_URI,
-               class_mappings=None):
+               class_mappings=None,
+               propagate_pillar_data_to_reclass=False):
 
-    data = get_nodeinfo(storage_type, inventory_base_uri, nodes_uri,
-                        classes_uri, minion_id, class_mappings)
+    nodes_uri, classes_uri = path_mangler(inventory_base_uri,
+                                          nodes_uri, classes_uri)
+    storage = get_storage(storage_type, nodes_uri, classes_uri,
+                          default_environment='base')
+    input_data = None
+    if propagate_pillar_data_to_reclass:
+        input_data = pillar
+    reclass = Core(storage, class_mappings, input_data=input_data)
+
+    data = reclass.nodeinfo(minion_id)
     params = data.get('parameters', {})
     params['__reclass__'] = {}
+    params['__reclass__']['nodename'] = minion_id
     params['__reclass__']['applications'] = data['applications']
     params['__reclass__']['classes'] = data['classes']
+    params['__reclass__']['environment'] = data['environment']
     return params
 
 
@@ -37,26 +50,31 @@ def top(minion_id, storage_type=OPT_STORAGE_TYPE,
         classes_uri=OPT_CLASSES_URI,
         class_mappings=None):
 
-    env = 'base'
-    # TODO: node environments
+    nodes_uri, classes_uri = path_mangler(inventory_base_uri,
+                                          nodes_uri, classes_uri)
+    storage = get_storage(storage_type, nodes_uri, classes_uri,
+                          default_environment='base')
+    reclass = Core(storage, class_mappings, input_data=None)
 
     # if the minion_id is not None, then return just the applications for the
     # specific minion, otherwise return the entire top data (which we need for
     # CLI invocations of the adapter):
     if minion_id is not None:
-        data = get_nodeinfo(storage_type, inventory_base_uri, nodes_uri,
-                            classes_uri, minion_id, class_mappings)
+        data = reclass.nodeinfo(minion_id)
         applications = data.get('applications', [])
+        env = data['environment']
         return {env: applications}
 
     else:
-        data = get_inventory(storage_type, inventory_base_uri, nodes_uri,
-                             classes_uri, class_mappings)
+        data = reclass.inventory()
         nodes = {}
         for node_id, node_data in data['nodes'].iteritems():
-            nodes[node_id] = node_data['applications']
+            env = node_data['environment']
+            if env not in nodes:
+                nodes[env] = {}
+            nodes[env][node_id] = node_data['applications']
 
-        return {env: nodes}
+        return nodes
 
 
 def cli():